requirements.txt
python-dotenv==1.0.0
packaging==23.2
bcrypt==4.1.2
twilio==9.0.0
openai==1.10.0
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.15
langchain-openai==0.0.3
chainlit==1.0.100
tiktoken==0.5.2
pytz==2023.4
python-dateutil==2.8.2
tqdm==4.66.1
pymongo==4.6.1
faiss-cpu==1.7.4
googlemaps==4.10.0
langchain-mongodb==0.1.3
